requirements.txt
pytubefix==7.4.0
openai-whisper==20240930
ffmpeg==1.4
streamlit==1.39.0
python-dotenv==1.0.1
langchain==0.3.3
langchain-openai==0.2.2
langgraph==0.2.37